+++ lib/src/client/call.dart
@@ -14,10 +14,16 @@
 // limitations under the License.
 
 import 'dart:async';
+import 'dart:convert';
+import 'dart:developer';
+
+import 'package:grpc/src/generated/google/rpc/status.pb.dart';
+import 'package:meta/meta.dart';
+import 'package:protobuf/protobuf.dart';
 
 import '../shared/message.dart';
+import '../shared/profiler.dart';
 import '../shared/status.dart';
-
 import 'common.dart';
 import 'connection.dart';
 import 'method.dart';
@@ -125,17 +131,24 @@ class WebCallOptions extends CallOptions {
   @override
   CallOptions mergedWith(CallOptions other) {
     if (other == null) return this;
-    if (other is! WebCallOptions) return super.mergedWith(other);
+
+    final mergedMetadata = Map.from(metadata)..addAll(other.metadata);
+    final mergedTimeout = other.timeout ?? timeout;
+    final mergedProviders = List.from(metadataProviders)
+      ..addAll(other.metadataProviders);
+
+    if (other is! WebCallOptions) {
+      return WebCallOptions._(Map.unmodifiable(mergedMetadata), mergedTimeout,
+          List.unmodifiable(mergedProviders),
+          bypassCorsPreflight: bypassCorsPreflight,
+          withCredentials: withCredentials);
+    }
 
     final otherOptions = other as WebCallOptions;
     final mergedBypassCorsPreflight =
         otherOptions.bypassCorsPreflight ?? bypassCorsPreflight;
     final mergedWithCredentials =
         otherOptions.withCredentials ?? withCredentials;
-    final mergedMetadata = Map.from(metadata)..addAll(otherOptions.metadata);
-    final mergedTimeout = otherOptions.timeout ?? timeout;
-    final mergedProviders = List.from(metadataProviders)
-      ..addAll(otherOptions.metadataProviders);
     return WebCallOptions._(Map.unmodifiable(mergedMetadata), mergedTimeout,
         List.unmodifiable(mergedProviders),
         bypassCorsPreflight: mergedBypassCorsPreflight,
@@ -163,7 +176,15 @@ class ClientCall<Q, R> implements Response {
   bool isCancelled = false;
   Timer _timeoutTimer;
 
-  ClientCall(this._method, this._requests, this.options) {
+  final TimelineTask _requestTimeline;
+  TimelineTask _responseTimeline;
+
+  ClientCall(this._method, this._requests, this.options,
+      [this._requestTimeline]) {
+    _requestTimeline?.start('gRPC Request: ${_method.path}', arguments: {
+      'method': _method.path,
+      'timeout': options?.timeout?.toString(),
+    });
     _responses = StreamController(onListen: _onResponseListen);
     if (options.timeout != null) {
       _timeoutTimer = Timer(options.timeout, _onTimedOut);
@@ -175,6 +196,7 @@ class ClientCall<Q, R> implements Response {
   }
 
   void _terminateWithError(GrpcError error) {
+    _finishTimelineWithError(error, _requestTimeline);
     if (!_responses.isClosed) {
       _responses.addError(error);
     }
@@ -229,8 +251,17 @@ class ClientCall<Q, R> implements Response {
       _terminateWithError(GrpcError.unavailable('Error making call: $e'));
       return;
     }
+    _requestTimeline?.instant('Request sent', arguments: {
+      'metadata': metadata,
+    });
     _requestSubscription = _requests
-        .map(_method.requestSerializer)
+        .map((data) {
+          _requestTimeline?.instant('Data sent', arguments: {
+            'data': data.toString(),
+          });
+          _requestTimeline?.finish();
+          return _method.requestSerializer(data);
+        })
         .handleError(_onRequestError)
         .listen(_stream.outgoingMessages.add,
             onError: _stream.outgoingMessages.addError,
@@ -241,8 +272,16 @@ class ClientCall<Q, R> implements Response {
     _onResponseListen();
   }
 
+  void _finishTimelineWithError(GrpcError error, TimelineTask timeline) {
+    timeline?.finish(arguments: {
+      'error': error.toString(),
+    });
+  }
+
   void _onTimedOut() {
-    _responses.addError(GrpcError.deadlineExceeded('Deadline exceeded'));
+    final error = GrpcError.deadlineExceeded('Deadline exceeded');
+    _finishTimelineWithError(error, _requestTimeline);
+    _responses.addError(error);
     _safeTerminate();
   }
 
@@ -261,18 +300,37 @@ class ClientCall<Q, R> implements Response {
       }
       _responses.onPause = _responseSubscription.pause;
       _responses.onResume = _responseSubscription.resume;
-      _responses.onCancel = _responseSubscription.cancel;
+      _responses.onCancel = cancel;
     }
   }
 
   /// Emit an error response to the user, and tear down this call.
   void _responseError(GrpcError error, [StackTrace stackTrace]) {
+    _finishTimelineWithError(error, _responseTimeline);
     _responses.addError(error, stackTrace);
     _timeoutTimer?.cancel();
     _requestSubscription?.cancel();
     _responseSubscription.cancel();
     _responses.close();
     _stream.terminate();
+  }
+
+  /// If there's an error status then process it as a response error
+  void _checkForErrorStatus(Map<String, String> metadata) {
+    final status = metadata['grpc-status'];
+    final statusCode = int.parse(status ?? '0');
+
+    if (statusCode != 0) {
+      final message = metadata['grpc-message'] == null
+          ? null
+          : Uri.decodeFull(metadata['grpc-message']);
+
+      _responseError(GrpcError.custom(
+        statusCode,
+        message,
+        decodeStatusDetails(metadata['grpc-status-details-bin']),
+      ));
+    }
   }
 
   /// Data handler for responses coming from the server. Handles header/trailer
@@ -288,15 +346,26 @@ class ClientCall<Q, R> implements Response {
         return;
       }
       try {
-        _responses.add(_method.responseDeserializer(data.data));
+        final decodedData = _method.responseDeserializer(data.data);
+        _responseTimeline?.instant('Data received', arguments: {
+          'data': decodedData.toString(),
+        });
+        _responses.add(decodedData);
         _hasReceivedResponses = true;
       } catch (e, s) {
         _responseError(GrpcError.dataLoss('Error parsing response'), s);
       }
     } else if (data is GrpcMetadata) {
       if (!_headers.isCompleted) {
-        // TODO(jakobr): Parse, and extract common headers.
         _headerMetadata = data.metadata;
+        if (_requestTimeline != null) {
+          _responseTimeline = timelineTaskFactory(
+              parent: _requestTimeline, filterKey: clientTimelineFilterKey);
+        }
+        _responseTimeline?.start('gRPC Response');
+        _responseTimeline?.instant('Metadata received', arguments: {
+          'headers': _headerMetadata.toString(),
+        });
         _headers.complete(_headerMetadata);
         return;
       }
@@ -305,17 +374,13 @@ class ClientCall<Q, R> implements Response {
         return;
       }
       final metadata = data.metadata;
+      _responseTimeline?.instant('Metadata received', arguments: {
+        'trailers': metadata.toString(),
+      });
       _trailers.complete(metadata);
-      // TODO(jakobr): Parse more!
-      if (metadata.containsKey('grpc-status')) {
-        final status = int.parse(metadata['grpc-status']);
-        final message = metadata['grpc-message'] == null
-            ? null
-            : Uri.decodeFull(metadata['grpc-message']);
-        if (status != 0) {
-          _responseError(GrpcError.custom(status, message));
-        }
-      }
+
+      /// Process status error if necessary
+      _checkForErrorStatus(metadata);
     } else {
       _responseError(GrpcError.unimplemented('Unexpected frame received'));
     }
@@ -348,16 +413,11 @@ class ClientCall<Q, R> implements Response {
       // Only received a header frame and no data frames, so the header
       // should contain "trailers" as well (Trailers-Only).
       _trailers.complete(_headerMetadata);
-      final status = _headerMetadata['grpc-status'];
-      // If status code is missing, we must treat it as '0'. As in 'success'.
-      final statusCode = status != null ? int.parse(status) : 0;
-      if (statusCode != 0) {
-        final message = _headerMetadata['grpc-message'] == null
-            ? null
-            : Uri.decodeFull(_headerMetadata['grpc-message']);
-        _responseError(GrpcError.custom(statusCode, message));
-      }
+
+      /// Process status error if necessary
+      _checkForErrorStatus(_headerMetadata);
     }
+    _responseTimeline?.finish();
     _timeoutTimer?.cancel();
     _responses.close();
     _responseSubscription.cancel();
@@ -371,6 +431,7 @@ class ClientCall<Q, R> implements Response {
       error = GrpcError.unknown(error.toString());
     }
 
+    _finishTimelineWithError(error, _requestTimeline);
     _responses.addError(error, stackTrace);
     _timeoutTimer?.cancel();
     _responses.close();
@@ -390,7 +451,9 @@ class ClientCall<Q, R> implements Response {
   @override
   Future<void> cancel() {
     if (!_responses.isClosed) {
-      _responses.addError(GrpcError.cancelled('Cancelled by client.'));
+      final error = GrpcError.cancelled('Cancelled by client.');
+      _responses.addError(error);
+      _finishTimelineWithError(error, _requestTimeline);
     }
     return _terminate();
   }
@@ -417,5 +480,25 @@ class ClientCall<Q, R> implements Response {
     try {
       await _terminate();
     } catch (_) {}
+  }
+}
+
+/// Given a string of base64url data, attempt to parse a Status object from it.
+/// Once parsed, it will then map each detail item and attempt to parse it into
+/// its respective GeneratedMessage type, returning the list of parsed detail items
+/// as a `List<GeneratedMessage>`.
+///
+/// Prior to creating the Status object we pad the data to ensure its length is
+/// an even multiple of 4, which is a requirement in Dart when decoding base64url data.
+///
+/// If any errors are thrown during decoding/parsing, it will return an empty list.
+@visibleForTesting
+List<GeneratedMessage> decodeStatusDetails(String data) {
+  try {
+    final parsedStatus = Status.fromBuffer(
+        base64Url.decode(data.padRight((data.length + 3) & ~3, '=')));
+    return parsedStatus.details.map(parseErrorDetailsFromAny).toList();
+  } catch (e) {
+    return <GeneratedMessage>[];
   }
 }