+++ lib/src/server/server.dart
@@ -77,25 +77,65 @@ class ServerTlsCredentials extends ServerCredentials {
   bool validateClient(Socket socket) => true;
 }
 
-/// A gRPC server.
+/// A gRPC server that serves via provided [ServerTransportConnection]s.
 ///
-/// Listens for incoming RPCs, dispatching them to the right [Service] handler.
-class Server {
+/// Unlike [Server], the caller has the responsibility of configuring and
+/// managing the connection from a client.
+class ConnectionServer {
   final Map<String, Service> _services = {};
   final List<Interceptor> _interceptors;
 
-  ServerSocket _insecureServer;
-  SecureServerSocket _secureServer;
   final _connections = <ServerTransportConnection>[];
 
   /// Create a server for the given [services].
-  Server(List<Service> services,
+  ConnectionServer(List<Service> services,
       [List<Interceptor> interceptors = const <Interceptor>[]])
       : _interceptors = interceptors {
     for (final service in services) {
       _services[service.$name] = service;
     }
   }
+
+  Service lookupService(String service) => _services[service];
+
+  Future<void> serveConnection(ServerTransportConnection connection) async {
+    _connections.add(connection);
+    ServerHandler_ handler;
+    // TODO(jakobr): Set active state handlers, close connection after idle
+    // timeout.
+    connection.incomingStreams.listen((stream) {
+      handler = serveStream_(stream);
+    }, onError: (error, stackTrace) {
+      if (error is Error) {
+        Zone.current.handleUncaughtError(error, stackTrace);
+      }
+    }, onDone: () {
+      // TODO(sigurdm): This is not correct behavior in the presence of
+      // half-closed tcp streams.
+      // Half-closed  streams seems to not be fully supported by package:http2.
+      // https://github.com/dart-lang/http2/issues/42
+      handler?.cancel();
+      _connections.remove(connection);
+    });
+  }
+
+  @visibleForTesting
+  ServerHandler_ serveStream_(ServerTransportStream stream) {
+    return ServerHandler_(lookupService, stream, _interceptors)..handle();
+  }
+}
+
+/// A gRPC server.
+///
+/// Listens for incoming RPCs, dispatching them to the right [Service] handler.
+class Server extends ConnectionServer {
+  ServerSocket _insecureServer;
+  SecureServerSocket _secureServer;
+
+  /// Create a server for the given [services].
+  Server(List<Service> services,
+      [List<Interceptor> interceptors = const <Interceptor>[]])
+      : super(services, interceptors);
 
   /// The port that the server is listening on, or `null` if the server is not
   /// active.
@@ -107,6 +147,9 @@ class Server {
 
   Service lookupService(String service) => _services[service];
 
+  /// Starts the [Server] with the given options.
+  /// [address] can be either a [String] or an [InternetAddress], in the latter
+  /// case it can be a Unix Domain Socket address.
   Future<void> serve(
       {dynamic address,
       int port,
@@ -135,35 +178,17 @@ class Server {
     }
     server.listen((socket) {
       // Don't wait for io buffers to fill up before sending requests.
-      socket.setOption(SocketOption.tcpNoDelay, true);
+      if (socket.address.type != InternetAddressType.unix) {
+        socket.setOption(SocketOption.tcpNoDelay, true);
+      }
       final connection = ServerTransportConnection.viaSocket(socket,
           settings: http2ServerSettings);
-      _connections.add(connection);
-      if (security != null && !security.validateClient(socket)) {
-        _printSocketError(
-            'cannot serve $address:$port - unable to validate client socket');
-        return socket.close();
+      serveConnection(connection);
+    }, onError: (error, stackTrace) {
+      if (error is Error) {
+        Zone.current.handleUncaughtError(error, stackTrace);
       }
-      ServerHandler_ handler;
-      // TODO(jakobr): Set active state handlers, close connection after idle
-      // timeout.
-      connection.incomingStreams.listen((stream) {
-        handler = serveStream_(stream);
-      }, onError: (error) {
-        print('Connection error: $error');
-      }, onDone: () {
-        // TODO(sigurdm): This is not correct behavior in the presence of
-        // half-closed tcp streams.
-        // Half-closed  streams seems to not be fully supported by package:http2.
-        // https://github.com/dart-lang/http2/issues/42
-        handler?.cancel();
-        _connections.remove(connection);
-      });
-    }, onError: _printSocketError);
-  }
-
-  void _printSocketError(Object error) {
-    print('Socket error: $error');
+    });
   }
 
   @visibleForTesting