+++ test/round_trip_test.dart
@@ -7,6 +7,7 @@ import 'package:grpc/src/client/channel.dart' hide ClientChannel;
 import 'package:grpc/src/client/connection.dart';
 import 'package:grpc/src/client/http2_connection.dart';
 import 'package:test/test.dart';
+import 'common.dart';
 
 class TestClient extends Client {
   static final _$stream = ClientMethod<int, int>('/test.TestService/stream',
@@ -14,9 +15,8 @@ class TestClient extends Client {
 
   TestClient(api.ClientChannel channel) : super(channel);
   ResponseStream<int> stream(int request, {CallOptions options}) {
-    final call =
-        $createCall(_$stream, Stream.fromIterable([request]), options: options);
-    return ResponseStream(call);
+    return $createStreamingCall(_$stream, Stream.value(request),
+        options: options);
   }
 }
 
@@ -28,10 +28,21 @@ class TestService extends Service {
         (List<int> value) => value[0], (int value) => [value]));
   }
 
+  static const requestFiniteStream = 1;
+  static const requestInfiniteStream = 2;
+
   Stream<int> stream(ServiceCall call, Future request) async* {
-    yield 1;
-    yield 2;
-    yield 3;
+    final isInfinite = 2 == await request;
+    for (var i = 1; i <= 3 || isInfinite; i++) {
+      yield i;
+      await Future.delayed(Duration(milliseconds: 100));
+    }
+  }
+}
+
+class TestServiceWithOnMetadataException extends TestService {
+  void $onMetadata(ServiceCall context) {
+    throw Exception('business exception');
   }
 }
 
@@ -46,7 +57,63 @@ class FixedConnectionClientChannel extends ClientChannelBase {
 }
 
 main() async {
-  test('round trip insecure connection', () async {
+  testTcpAndUds('round trip insecure connection', (address) async {
+    // round trip test of insecure connection.
+    final Server server = Server([TestService()]);
+    await server.serve(address: address, port: 0);
+
+    final channel = FixedConnectionClientChannel(Http2ClientConnection(
+      address,
+      server.port,
+      ChannelOptions(credentials: ChannelCredentials.insecure()),
+    ));
+    final testClient = TestClient(channel);
+    expect(await testClient.stream(TestService.requestFiniteStream).toList(),
+        [1, 2, 3]);
+    server.shutdown();
+  });
+
+  testTcpAndUds('round trip secure connection', (address) async {
+    // round trip test of secure connection.
+    final Server server = Server([TestService()]);
+    await server.serve(
+        address: address,
+        port: 0,
+        security: ServerTlsCredentials(
+            certificate: File('test/data/localhost.crt').readAsBytesSync(),
+            privateKey: File('test/data/localhost.key').readAsBytesSync()));
+
+    final channel = FixedConnectionClientChannel(Http2ClientConnection(
+      address,
+      server.port,
+      ChannelOptions(
+          credentials: ChannelCredentials.secure(
+              certificates: File('test/data/localhost.crt').readAsBytesSync(),
+              authority: 'localhost')),
+    ));
+    final testClient = TestClient(channel);
+    expect(await testClient.stream(TestService.requestFiniteStream).toList(),
+        [1, 2, 3]);
+    server.shutdown();
+  });
+
+  test('exception in onMetadataException', () async {
+    final Server server = Server([TestServiceWithOnMetadataException()]);
+    await server.serve(address: 'localhost', port: 0);
+
+    final channel = FixedConnectionClientChannel(Http2ClientConnection(
+      'localhost',
+      server.port,
+      ChannelOptions(credentials: ChannelCredentials.insecure()),
+    ));
+    final testClient = TestClient(channel);
+    await expectLater(
+        testClient.stream(TestService.requestFiniteStream).toList(),
+        throwsA(isA<GrpcError>()));
+    await server.shutdown();
+  });
+
+  test('cancellation of streaming subscription propagates properly', () async {
     final Server server = Server([TestService()]);
     await server.serve(address: 'localhost', port: 0);
 
@@ -56,29 +123,8 @@ main() async {
       ChannelOptions(credentials: ChannelCredentials.insecure()),
     ));
     final testClient = TestClient(channel);
-    expect(await testClient.stream(1).toList(), [1, 2, 3]);
-    server.shutdown();
-  });
-
-  test('round trip secure connection', () async {
-    final Server server = Server([TestService()]);
-    await server.serve(
-        address: 'localhost',
-        port: 0,
-        security: ServerTlsCredentials(
-            certificate: File('test/data/localhost.crt').readAsBytesSync(),
-            privateKey: File('test/data/localhost.key').readAsBytesSync()));
-
-    final channel = FixedConnectionClientChannel(Http2ClientConnection(
-      'localhost',
-      server.port,
-      ChannelOptions(
-          credentials: ChannelCredentials.secure(
-              certificates: File('test/data/localhost.crt').readAsBytesSync(),
-              authority: 'localhost')),
-    ));
-    final testClient = TestClient(channel);
-    expect(await testClient.stream(1).toList(), [1, 2, 3]);
-    server.shutdown();
+    expect(await testClient.stream(TestService.requestInfiniteStream).first, 1);
+    await channel.shutdown();
+    await server.shutdown();
   });
 }